// data/github/current/ui/packages/notifications-inbox/notifications-inbox-v1/components-v1/list/__generated__/InboxList_v1_threadFragment.graphql.ts
export type InboxList_v1_threadFragment$data = {
  readonly id: string;
  readonly notificationThreads: {
    readonly edges: ReadonlyArray<{
      readonly cursor: string;
      readonly node: {
        readonly id: string;
        readonly isDone: boolean;
        readonly isUnread: boolean;
        readonly list: {
          readonly __typename: "Enterprise";
          readonly slug: string;
        } | {
          readonly __typename: "Organization";
          readonly login: string;
        } | {
          readonly __typename: "Repository";
          readonly name: string;
          readonly owner: {
            readonly login: string;
          };
        } | {
          readonly __typename: "Team";
          readonly name: string;
          readonly organization: {
            readonly login: string;
          };
          readonly slug: string;
        } | {
          readonly __typename: "User";
          readonly login: string;
        } | {
          // This will never be '%other', but we need some
          // value in case none of the concrete values match.
          readonly __typename: "%other";
        };
        readonly subject: {
          readonly __typename: "Commit";
          readonly id: string;
        } | {
          readonly __typename: "Discussion";
          readonly id: string;
          readonly number: number;
        } | {
          readonly __typename: "Issue";
          readonly id: string;
          readonly number: number;
        } | {
          readonly __typename: "PullRequest";
          readonly id: string;
          readonly number: number;
        } | {
          readonly __typename: "TeamDiscussion";
          readonly id: string;
        } | {
          // This will never be '%other', but we need some
          // value in case none of the concrete values match.
          readonly __typename: "%other";
        };
        readonly url: string;
        readonly " $fragmentSpreads": FragmentRefs<"InboxListRow_v1_fragment">;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  };
  readonly " $fragmentType": "InboxList_v1_threadFragment";
};
export type InboxList_v1_threadFragment$key = {
  readonly " $data"?: InboxList_v1_threadFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"InboxList_v1_threadFragment">;
};