// data/github/current/ui/packages/notifications-inbox/notifications-inbox-v1/components-v1/list/__generated__/InboxListV1Query.graphql.ts
export type InboxListV1Query$variables = {
  cursor?: string | null | undefined;
  first?: number | null | undefined;
  id: string;
  query: string;
};
export type InboxListV1Query$data = {
  readonly node: {
    readonly " $fragmentSpreads": FragmentRefs<"InboxList_v1_fragment">;
  } | null | undefined;
};
export type InboxListV1Query = {
  response: InboxListV1Query$data;
  variables: InboxListV1Query$variables;
};